Insulator Detection using Faster-RCNN model/requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
setuptools==75.2.0
six==1.16.0
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
typing_extensions==4.12.2
